Initial Commit
Build with webpack, all dependencies are in package.json. The only dependency for deployment is react.
Most of the 'framework' is here, but will need to be modified as development progresses. Currently the interface is exclusively debug components for developing the underlying store, router, and AudioInterface architecture. Only uses the backend for audio normilization.