Many changes mainly surrounding the Store.
Mutations may now be batched using the `will_update` and `did_update` functions. This is very handy for scenarios in which the display (reflective of the state) needs to be updated many times, but the only 'real' mutation of the state is done after that interaction (eg, dragging the position of an object). This allows for undo and redo to be managed with a finer grain.

Many of the components have been updated to utilize this change, or additionally remove unnecessary re-renders. This has greatly improved the performance of the editor.

General cleanup and style updates are included as well.